# app/openapi.yaml
openapi: 3.0.0
info:
  title: Sample API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9
servers:
  - url: /api
paths:
...
  /fitbit/profile:
    get:
      tags:
        - Fitbit API wrapper
      summary: Login user with Fitbit. You must provide an Authorization Bearer token.
      operationId: app.views.fitbit_wrapper.profile
      responses:
        '200':
          description: 'Login successful.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: User not authorized.
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      x-basicInfoFunc: app.views.auth.basic_auth
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: app.views.auth.jwt_auth
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
          example: test_user
        display_name:
          type: string
          example: Test User
        last_login_time:
          type: string
          example: 2020-03-29T15:32:13Z
...